slack-api/requirements.txt
slack-sdk==3.26.2
python-dotenv==1.0.0